Add support for downloading batches of samples (#415)
Added `download_samples()` function and `onecodex download samples` CLI command for downloading batches of samples as FASTA/Q files:
- All samples in your account
- Samples optionally filtered by project name or ID
- Samples optionally filtered by tag names

The code was ported from: https://gist.github.com/clausmith/c5da3145412265014151cfcad5493180

Closes DEV-6563

+++ onecodex/models/helpers.py
@@ -1,6 +1,7 @@
 import click
 import inspect
 import os
+import os.path
 import requests
 
 from onecodex.exceptions import OneCodexException, UnboundObject
@@ -156,7 +157,7 @@ class ResourceDownloadMixin(object):
                 path = os.path.join(os.getcwd(), _filename)
 
             if path and os.path.exists(path):
-                raise OneCodexException("{} already exists! Will not overwrite.".format(path))
+                raise OneCodexException("{} already exists. Will not overwrite.".format(path))
 
             if use_potion_session:
                 session = self._resource._client.session
@@ -191,7 +192,7 @@ class ResourceDownloadMixin(object):
                     for data in resp.iter_content(chunk_size=1024):
                         f_out.write(data)
         except KeyboardInterrupt:
-            if path:
+            if path and os.path.exists(path):
                 os.remove(path)
             raise
         except requests.exceptions.HTTPError as exc:
@@ -200,7 +201,8 @@ class ResourceDownloadMixin(object):
             elif exc.response.status_code == 402:
                 raise OneCodexException(
                     "You must either have a premium platform account or be in "
-                    "a notebook environment to download files."
+                    "a notebook environment to download files. Please feel free to contact us "
+                    "about your subscription at support@onecodex.com."
                 )
             elif exc.response.status_code == 403:
                 raise OneCodexException("You are not authorized to download this file.")
